Visual Studio 2005 rules script for ANTLR3. Requires a small change from Ter on where the outputs go when the inputs are a sub-directory or visual studio and other IDEs will not be able to track the output files. Also, I have adopted a new file suffix convention that will allow rules such as this, ant and gmake to predict the output files:
*.g3/*g3pl - Assumed to be a Parser/Lexer combination;
*.g3p  - Assumed to be a parser grammar;
*.g3l - assumed to be a lexer grammar;
*.g3t - assumed to be a tree grammar;

Of course, nobody is strictly bound to this convention but I will try it on the group and see what people say. I also belive that we should always produce outputs based upon th efile name and not the grammar name, which would be more for documentation - need to discuss with Ter and see if he is convinced.

[git-p4: depot-paths = "//depot/code/antlr/main/": change = 2919]