Preparation for Maven build.:
The ANTLR 2 tool for MAven expects that aall interdependent v2 grammars are in the same directory. This
seems reasonable, as V2 does not support specifiying that the vocabulary files for the tokens are
in a different directory. Hence the v2 and v3 grammar files are now collected together in the hierarchy
which means there is no longer a need to copy around the token text files.

This check in also adds the initial Maven project defintion, which will change over the next few check-ins.

[git-p4: depot-paths = "//depot/code/antlr/main/": change = 5364]