Lots of optimizations (though the next one to be checked in is the huge win) for AST building and vector factories.
A large part of tree rewriting was the creation of vectors to hold AST nodes. Although I had created a vector factory, for some reason I never got around to creating a proper one, that pre-allocated the vectors in chunks and so on. I guess I just forgot to. Hence a big win here is prevention of calling malloc lots and lots of times to create vectors.

A second inprovement was to change teh vector definition such that it holds a certain number of elements wihtin the vector structure itself, rather than malloc and freeing these. Currently this is set to 8, but may increase. For AST construction, this is generally a big win because AST nodes don't often have many individual children unless there has not been any shaping going on in the parser. But if you are not shaping, then you don't really need a tree.

Other perforamnce inprovements here include not calling functions indirectly within token stream and common token stream. Hence tokens are claimed directly from the vectors. Users can override these funcitons of course and all this means is that if you override tokenstreams then you pretty much have to provide all the mehtods, but then I think you woudl have to anyway (and I don't know of anyone that has wanted to do this as you can carry your own structure around with the tokens anyway and that is much easier).

[git-p4: depot-paths = "//depot/code/antlr/main/": change = 5558]