C Runtime - Further Optimizations
Within grammars that used scopes and were intended to parse large inputs with many rule nests,
the creation anf deletion of the scopes themselves became significant. Careful analysis shows that
for most grammars, while a parse could create and delete 20,000 scopes, the maxium depth of
any scope was only 8. 

This change therefore changes the scope implementation so that it does not free scope memory when
it is popped but just tracks it in a C runtime stack, eventually freeing it when the stack is freed. This change
caused the allocation of only 12 scope structures instead of 20,000 for the extreme example case.

This change means that scope users must be carefule (as ever in C) to initializae their scope elements
correctly as:

1) If not you may inherit values from a prior use of the scope structure;
2) SCope structure are now allocated with malloc and not calloc;

Also, when using a custom free function to clean a scope when it is popped, it is probably a good idea
to set any free'd pointers to NULL (this is generally good C programmig practice in any case)

[git-p4: depot-paths = "//depot/code/antlr/main/": change = 5567]