C Runtime - Vastly improve teh reuse of allocated memory for nodes in tree rewriting.
A problem for all targets at the moment si that the rewrite logic generated by ANTLRR makes no attempt
to reuse any resources, it merely gurantees that the tree shape at the end is correct. To some extent this
is mitigated by the garbage collection systems of Java and .Net, even thoguh it is still an overhead to
keep creating so many modes.

This change implements the first of two C runtime changes that make best efforst to track when a node
has become orphaned and will never be reused, based on inherent knowledge of the rewrite logic (which in
the long term is not a great soloution). 

Much of the rewrite logic consists of creating a niilnode into which child nodes are appended. At: rulePost
processing time; when a rewrite stream is closed; and when becomeRoot is called, there are many situations
where the root of the tree that will be manipulted, or is finished with (in the case of rewrtie streams), where the
nilNode was just a temporary creation for the sake of the rewrite itself. 

In these cases we can see that the nilNode would just be left ot rot in the node factory that tracks all the tree nodes.
Rather than leave these in the factory to rot, we now keep a resuse stck and always reuse any node on this
stack before claimin a new node from the factory pool.

This single change alone reduces memory usage in the test case (20,604 line C program and a GNU C parser) from nearly a GB, to 276MB. This is still way more memory than we shoudl need to do this operation, even on such a large input file,
but the reduction results in a huge performance increase and greatly reduced system time spent on allocations.

After this optimizatoin, comparison with gcc yeilds:

time gcc -S a.c
a.c:1026: warning: conflicting types for built-in function ‘vsprintf’
a.c:1030: warning: conflicting types for built-in function ‘vsnprintf’
a.c:1041: warning: conflicting types for built-in function ‘vsscanf’
0.21user 0.01system 0:00.22elapsed 97%CPU (0avgtext+0avgdata 0maxresident)k
0inputs+240outputs (0major+8345minor)pagefaults 0swaps

and

time ./jimi
Reading a.c
0.28user 0.11system 0:00.39elapsed 98%CPU (0avgtext+0avgdata 0maxresident)k
0inputs+0outputs (0major+66609minor)pagefaults 0swaps

And we can now interpolate the fact that the only major differnce is now the huge disparity in memory allocations. A
future optimization of vector pooling, to sepate node resue from vector reuse, currently looks promising for further
reuse of memory.

Finally, a static analysis of the rewrte code, plus a realtime analysis of the heap at runtime, may
well give us a reasonable memory usage pattern. In reality though, it is the generated rewrite logic
that must becom optional at not continuously rewriting things that it need not, as it ascends the
rule chain.

[git-p4: depot-paths = "//depot/code/antlr/main/": change = 5575]