Rename/Move: From: //depot/code/antlr/main/src/org/antlr/grammar/v3/... To: //depot/code/antlr/main/src/main/antlr/org/antlr/grammar/v3/...
Move the v3 grammar files for the ANTLR tool to reflect the Maven standards. This is part of a gradual move to be 100% Maven like for the builds.

Thsi change also updates the pom to use the correct version of the antlr plugin, which is able to correctly predict the order in which .g files should be parsed by the ANTLR tool

[git-p4: depot-paths = "//depot/code/antlr/main/": change = 5616]