Rename/Move: From: //depot/code/antlr/main/src/org/antlr/codegen/templates/... To: //depot/code/antlr/main/src/main/resources/org/antlr/codegen/templates/...
Relocate the code generating templates to exist in the directory set that maven expects.

When checking in your templates, you may find it easiest to make a copy of what you have, revert the change in perforce, then just check out the template in the new location, and copy the changes back over. Nobody has oore than two files open at the moment.

[git-p4: depot-paths = "//depot/code/antlr/main/": change = 5620]