ANTLR C Runtime - Add a topological sort class to the ANTLR runtime
A useful utlitiy in code genereation, optimization and so on is the abilitly to sort topologically, 
whereby elements are reordered such that elements that depend on other elements are
sorted to appera in a list after the things they depend upon.

This change iplements the ANTLR3_TOPO class, which can be used to create
a list of node numbers and edges (node 0 depends on node 6 is an edge: 0->6)
then sort the supplied nodes into topological order by depth first dependency analysis.

This algorithm is the depth first serach of edges listed in reverse post order, which gives
the desired ordering (or at aleast one of possibly many valid orderings). It is similat
to the one used in the ANTLR tool except that this version detects cycles such as:

0->4->6->9->0

and will abort processing if it detects a cycle, returning the cycle to the caller to do
with as they will. It can also be used to sort an ANTLR3_VECTOR in place, which is
of course the structure that stores child nodes in an AST. This means that it can be
called upon to sort in place a list of children in an AST node, with obvious applications
for tree walking.

[git-p4: depot-paths = "//depot/code/antlr/main/": change = 6029]